(SM-1) Hello World Component added.
