+++ src/components/App.test.tsx
@@ -4,6 +4,6 @@ import App from './App';
 
 test('renders learn react link', () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  const linkElement = screen.getByText(/HELLO WORLD/i);
   expect(linkElement).toBeInTheDocument();
 });
